Moved routes folder for cleaner code
